Ignores "R0903: Too few public methods".
main.py uses QMainWindow which requires it to be in a class by itself. Linter suggests not using a class since there aren't any public methods (Something something... objects shouldn't be used just for data). However, this is better for maintainability if we ever need to add more methods or override QMainWindow methods (It's also recommended to do it this way).