Implement logic for different types of file lists.
Logic such as there is a combination of shows and movies in a list (not allowed), there are only episodes but multiple tv show series (not allowed), there are only episodes and there is only one show (allowed), and there are movie files (allowed).

Additional checking, e.g., whether a movie MediaRecord has a year or title attribute will be down in backend.